+++ crates/config/src/schema/renderers/mod.rs
@@ -1,5 +1,8 @@
 #[cfg(feature = "json_schema")]
 pub mod json_schema;
 
+#[cfg(feature = "template")]
+pub mod template;
+
 #[cfg(feature = "typescript")]
 pub mod typescript;