+++ crates/macros/src/config_enum/variant.rs
@@ -1,6 +1,7 @@
 use crate::common::FieldSerdeArgs;
 use crate::utils::{
-    extract_comment, extract_common_attrs, extract_deprecated, format_case, map_option_quote,
+    extract_comment, extract_common_attrs, extract_deprecated, format_case, get_meta_path,
+    map_option_quote,
 };
 use darling::FromAttributes;
 use proc_macro2::{Ident, TokenStream};
@@ -17,6 +18,7 @@ pub struct VariantArgs {
 
 pub struct Variant<'l> {
     pub args: VariantArgs,
+    pub default: bool,
     pub serde_args: FieldSerdeArgs,
     pub attrs: Vec<&'l Attribute>,
     pub name: &'l Ident,
@@ -54,8 +56,13 @@ impl<'l> Variant<'l> {
             format_case(format, variant.ident.to_string().as_str(), true)
         };
 
+        let attrs = extract_common_attrs(&variant.attrs);
+
         Variant {
-            attrs: extract_common_attrs(&variant.attrs),
+            default: attrs
+                .iter()
+                .any(|v| get_meta_path(&v.meta).is_ident("default")),
+            attrs,
             name: &variant.ident,
             value,
             args,
@@ -102,7 +109,7 @@ impl<'l> Variant<'l> {
     }
 
     pub fn get_schema_type(&self) -> TokenStream {
-        let name = map_option_quote("name", Some(self.name.to_string()));
+        let name = self.name.to_string();
         let description = map_option_quote("description", extract_comment(&self.attrs));
         let deprecated = map_option_quote("deprecated", extract_deprecated(&self.attrs));
 
@@ -120,8 +127,8 @@ impl<'l> Variant<'l> {
 
         quote! {
             SchemaField {
+                name: #name.into(),
                 type_of: #type_of,
-                #name
                 #description
                 #deprecated
                 ..Default::default()