+++ crates/types/src/enums.rs
@@ -3,6 +3,7 @@ use crate::SchemaField;
 
 #[derive(Clone, Debug, Default)]
 pub struct EnumType {
+    pub default_index: Option<usize>,
     pub description: Option<String>,
     pub name: Option<String>,
     pub values: Vec<LiteralValue>,