+++ crates/types/src/unions.rs
@@ -9,10 +9,17 @@ pub enum UnionOperator {
 
 #[derive(Clone, Debug, Default)]
 pub struct UnionType {
+    pub default_index: Option<usize>,
     pub description: Option<String>,
     pub name: Option<String>,
     pub partial: bool,
     pub operator: UnionOperator,
     pub variants: Option<Vec<SchemaField>>,
     pub variants_types: Vec<Box<SchemaType>>,
+}
+
+impl UnionType {
+    pub fn is_nullable(&self) -> bool {
+        self.variants_types.iter().any(|v| v.is_null())
+    }
 }